Split into sub-module with a semantic/functionality meaning
